Add files via upload
Task 1

"Guess the Number" is an engaging and interactive game where the player attempts to guess a randomly selected number within a specified range. The game offers multiple rounds, allowing players to improve their guessing skills and achieve a higher score. The player has a limited number of attempts (lives) to guess the correct number in each round, adding a layer of challenge and excitement.